inference/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
addict==2.4.0
aiofiles==23.2.1
alabaster==0.7.13
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.3
attrs==23.2.0
Babel==2.14.0
black==24.2.0
boto3==1.34.42
botocore==1.34.42
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dill==0.3.8
docutils==0.20.1
exceptiongroup==1.2.0
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
flake8==7.0.0
flake8-docstrings==1.7.0
fonttools==4.48.1
fsspec==2024.2.0
gmplot==1.4.1
gradio==4.12.0
gradio_client==0.8.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mccabe==0.7.0
mdurl==0.1.2
motmetrics==1.4.0
mypy==1.8.0
mypy-extensions==1.0.0
nanoid==2.0.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
orjson==3.9.14
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
plyfile==1.0.3
psutil==5.9.8
pycocotools==2.0.7
pycodestyle==2.11.1
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.0.3
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.0
scalabel @ git+https://github.com/scalabel/scalabel.git@071d073598e7c988134dc70a5c0c52761226bf42
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.36.3
tabulate==0.9.0
terminaltables==3.1.10
tifffile==2023.7.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.27.1
websockets==11.0.3
xmltodict==0.13.0
yapf==0.40.2
zipp==3.17.0
